fix: add missing kuzu_ffi_test.rs and apply cargo fmt formatting
- Add tests/integration/kuzu_ffi_test.rs (was untracked, referenced in
  Cargo.toml but not committed, causing CI cargo fmt to error)
- Apply cargo fmt to fix formatting in memory/mod.rs, plugin.rs,
  update.rs, and cargo_lock_cxx_consistency_test.rs
