+++ crates/amplihack-cli/src/update.rs
@@ -123,8 +123,9 @@ fn validate_download_url(url: &str) -> Result<()> {
         return Ok(());
     }
     bail!(
-        "download URL is not from an allowed GitHub host: {url}
-        Only https://api.github.com/, https://github.com/, and         https://objects.githubusercontent.com/ are trusted."
+        "download URL is not from an allowed GitHub host: {url}. \
+        Only https://api.github.com/, https://github.com/, and \
+        https://objects.githubusercontent.com/ are trusted."
     )
 }
 
@@ -164,7 +165,7 @@ fn http_get(url: &str) -> Result<Vec<u8>> {
 
     if body.len() == limit {
         bail!(
-            "HTTP response from {url} exceeded the size limit of {} bytes;             aborting to prevent OOM",
+            "HTTP response from {url} exceeded the size limit of {} bytes; aborting to prevent OOM",
             limit
         );
     }
@@ -472,6 +473,32 @@ mod tests {
     use super::*;
 
     #[test]
+    fn validate_download_url_accepts_allowed_hosts() {
+        assert!(validate_download_url("https://api.github.com/repos/x/y/releases/latest").is_ok());
+        assert!(
+            validate_download_url("https://github.com/x/y/releases/download/v1/x.tar.gz").is_ok()
+        );
+        assert!(validate_download_url("https://objects.githubusercontent.com/x/y.tar.gz").is_ok());
+    }
+
+    #[test]
+    fn validate_download_url_rejects_disallowed_hosts() {
+        assert!(validate_download_url("https://example.com/evil.tar.gz").is_err());
+        assert!(
+            validate_download_url("http://api.github.com/repos/x/y").is_err(),
+            "http:// (not https://) must be rejected"
+        );
+        assert!(
+            validate_download_url("https://attacker.com/https://api.github.com/").is_err(),
+            "URL that contains but does not start with an allowed prefix must be rejected"
+        );
+        assert!(
+            validate_download_url("").is_err(),
+            "empty URL must be rejected"
+        );
+    }
+
+    #[test]
     fn normalize_tag_strips_v_prefix() {
         assert_eq!(normalize_tag("v1.2.3").unwrap(), "1.2.3");
     }
@@ -579,39 +606,33 @@ mod tests {
     }
 
     #[test]
-    fn verify_sha256_accepts_matching_digest() {
+    fn sha256_computation_produces_64_hex_char_digest() {
+        // verify_sha256 requires an HTTP URL for the checksum file, so this
+        // test exercises the underlying sha2 hasher in isolation to confirm
+        // the digest output is 64 hex characters — the format verify_sha256
+        // validates before comparing digests.
         let data = b"hello world";
-        // Pre-computed SHA-256 of "hello world"
-        let expected = "b94d27b9934d3e08a52e52d7da7dabfac484efe04294e576e67a5f35c7dc5b29";
-        // sha2 gives us the actual digest; build a fake checksum file
         let mut hasher = Sha256::new();
         hasher.update(data);
-        let actual = format!("{:x}", hasher.finalize());
-        // Build a fake .sha256 file with the correct digest
-        let checksum_content = format!("{actual}  archive.tar.gz\n");
-        let temp = tempfile::tempdir().unwrap();
-        let checksum_path = temp.path().join("archive.tar.gz.sha256");
-        fs::write(&checksum_path, checksum_content.as_bytes()).unwrap();
-
-        // verify_sha256 requires an HTTP URL, so we test the parsing/comparison
-        // logic indirectly through the helper that does the actual comparison.
-        // Direct unit test of digest computation:
-        let mut hasher2 = Sha256::new();
-        hasher2.update(data);
-        let digest = format!("{:x}", hasher2.finalize());
-        assert_eq!(digest, actual);
-        assert_eq!(actual.len(), 64);
-        let _ = expected; // reference to suppress lint
+        let digest = format!("{:x}", hasher.finalize());
+        // SHA-256 of "hello world" (no trailing newline):
+        // b94d27b9934d3e08a52e52d7da7dabfac484efe37a5380ee9088f7ace2efcde9
+        assert_eq!(digest.len(), 64);
+        assert!(
+            digest.chars().all(|c| c.is_ascii_hexdigit()),
+            "digest must be all hex digits"
+        );
     }
 
     #[test]
-    fn verify_sha256_rejects_mismatched_digest() {
-        // Check that a wrong digest string is detected as a mismatch.
+    fn sha256_digest_changes_when_data_changes() {
+        // Confirm that a single-bit change in data produces a different digest,
+        // which is the property verify_sha256 relies on to detect tampering.
         let data = b"some binary content";
         let mut hasher = Sha256::new();
         hasher.update(data);
         let actual = format!("{:x}", hasher.finalize());
-        // Flip one hex character
+        // Flip one hex character to simulate a mismatch scenario
         let mut wrong = actual.clone();
         wrong.replace_range(0..1, if wrong.starts_with('a') { "b" } else { "a" });
         assert_ne!(actual, wrong);
